+++ DigitalWorld/Assets/Logic/Scripts/Implements/Actions/Game/CreateCharacter.cs
@@ -39,7 +39,7 @@ namespace DigitalWorld.Logic.Actions.Game
         /// 如果是瞬时类行动 则建议忽视该函数 留空
         /// </summary>
         /// <param name="delta"></param>
-        protected override void OnUpdate(float delta)
+        protected override void OnUpdate(int delta)
         {
             base.OnUpdate(delta);
